backend/requirements.txt
google-genai
python-dotenv
beautifulsoup4
crawl4ai
faiss-cpu
numpy
rank_bm25
fastapi
uvicorn
requests
langchain
langchain-core
langchain-community
langchain-google-genai
pypdf
onnxruntime
tokenizers
transformers
langchain-text-splitters
langgraph
